initial commit
This commit introduces introduces the first of the deter switch drivers,
the snmp driver. It may be possible that this is the only driver that is
ever required  for conventional switches (holds breath), as most modern
switches (now including cumulus) support the Q-BRIDGE SNMP API for
modifying vlans.

The snmp driver is a generic snmp driver that uses the Q-BRIDGE SNMP
specification to control vlan configurations on a switch.  At this time
implements the setPortAccess and setPortTrunk commands as specified in
the deter functional architecture spec. It also provides port status
query capability and neighbor discovery through LLDP query over SNMP.